refactor: move ActorDto to actor package and add ActorDetailsDto
