+++ repository/src/main/java/com/baghdad/repository/model/actor/ActorDto.kt
@@ -1,4 +1,4 @@
-package com.baghdad.repository.model
+package com.baghdad.repository.model.actor
 
 data class ActorDto(
     val id : Long,
